LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_anchor_download_block_downloads.tentative-expected.txt

CONSOLE MESSAGE: Blocked script execution in 'about:srcdoc' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: Blocked script execution in 'about:srcdoc' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: Blocked script execution in 'about:srcdoc' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.


FAIL <a download> triggered download in sandbox is blocked. undefined is not an object (evaluating 'anchor.href = `support/download_stash.py?token=${download_token}` +
        `&finish-delay=${StreamDownloadFinishDelay()}`')
FAIL <a download> triggered download in sandbox is blocked before a request is made. undefined is not an object (evaluating 'anchor.href = `support/download_stash.py?token=${download_token}`')
FAIL <a > triggered download in sandbox is blocked. undefined is not an object (evaluating 'anchor.href = `support/download_stash.py?token=${download_token}` +
        `&finish-delay=${StreamDownloadFinishDelay()}`')
FAIL <a target="_blank" > triggered download in sandbox is blocked. assert_equals: Expect no download to happen, but got one. expected "TOKEN_NOT_SET" but got "TOKEN_SET"
FAIL <a target="_blank" rel="noopener" > triggered download in sandbox is blocked. assert_equals: Expect no download to happen, but got one. expected "TOKEN_NOT_SET" but got "TOKEN_SET"
